+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+0.0.8 / 2023-11-03
+------------------
+- Changed
+    - Fix parse xlsx without xl/styles.xml
+    - Refactoring and optimizations
+
 0.0.7 / 2023-10-27
 ------------------
 - Changed
-  - Fix parse xlsx without xl/sharedStrings.xml
-  - Fix parse cell with number and inlineStr data types
+    - Fix parse xlsx without xl/sharedStrings.xml
+    - Fix parse cell with number and inlineStr data types
 
 0.0.6 / 2023-09-16
 ------------------